chore: Screen, Component 패키지 분리
