+++ app/src/main/java/woowacourse/payments/ui/component/NewCardTopBar.kt
@@ -1,0 +1,56 @@
+package woowacourse.payments.ui.component
+
+import android.content.res.Configuration
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.Check
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import woowacourse.payments.ui.theme.AndroidpaymentsTheme
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun NewCardTopBar(
+    onBackClick: () -> Unit,
+    onSaveClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    TopAppBar(
+        title = { Text("카드 추가") },
+        navigationIcon = {
+            IconButton(onClick = { onBackClick() }) {
+                Icon(
+                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                    contentDescription = "뒤로 가기",
+                )
+            }
+        },
+        actions = {
+            IconButton(onClick = { onSaveClick() }) {
+                Icon(
+                    imageVector = Icons.Filled.Check,
+                    contentDescription = "완료",
+                )
+            }
+        },
+        modifier = modifier,
+    )
+}
+
+@Composable
+@Preview(name = "기본 상태", showBackground = true)
+@Preview(name = "다크 모드", uiMode = Configuration.UI_MODE_NIGHT_YES, showBackground = true)
+fun CardTopBarPreview() {
+    AndroidpaymentsTheme {
+        NewCardTopBar(
+            onBackClick = {},
+            onSaveClick = {},
+        )
+    }
+}
